Add update and token auth

+++ app/controllers/answers_controller.rb
@@ -1,19 +1,40 @@
 class AnswersController < ApplicationController
   def create
   answer = Answer.new(question_id: params[:question_id], member_id: params[:member_id], desc: params[:desc], accepted: params[:accepted])
-    if answer.save
-      render json: answer
+    if answer.member.token == params[:token]
+      if answer.save
+        render json: answer
+      else
+        render json: answer.errors
+      end
     else
-      render json: answer.errors
+      render json: "You shall not pass"
+    end
+  end
+
+  def update
+    answer = Answer.find(params[:id])
+    if answer.member.token == params[:token]
+      if answer.save
+        render json: answer
+      else
+        render json: answer.errors
+      end
+    else
+      render json: "You shall not pass"
     end
   end
 
   def destroy
     answer= Answer.find(params[:id])
-    if answer.destroy
-      render json: "ITS HAD BEEN DECIDED"
+    if answer.member.token == params[:token]
+      if answer.destroy
+        render json: "ITS HAD BEEN DECIDED"
+      else
+        render json: answer.errors
+      end
     else
-      render json: answer.errors
+      render json: "You shall not pass"
     end
   end
